+++ back-end/src/main/java/com/tdd/backend/post/service/SharingService.java
@@ -1,7 +1,8 @@
-package com.tdd.backend.post;
+package com.tdd.backend.post.service;
 
 import org.springframework.stereotype.Service;
 
+import com.tdd.backend.post.PostRepository;
 import com.tdd.backend.post.data.SharingDto;
 
 import lombok.RequiredArgsConstructor;
